requirements-test.txt
# Test requirements for Utility Tariff integration
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-homeassistant-custom-component==0.13.171
pytest-cov==5.0.0
pytest-timeout==2.3.1
aiohttp==3.10.11
PyPDF2==3.0.1